# [TEST] Adds performance test for vectorised global alingment.