Resolves #1574 

Adds benchmarks for vectorised global alingment.
Here are the results from my local benchmark with SSE4.

```console
Running ./alignment/global_affine_alignment_simd_benchmark
Run on (8 X 2600 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
Load Average: 2.92, 2.99, 3.08
----------------------------------------------------------------------------------------------------------------------
Benchmark                                                            Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------------------------------
seqan3_affine_dna4_simd<score>/0/real_time                     7600672 ns      7558451 ns           91 CUPS=3.07186G/s cells=23.3482M total=-6.47574M
seqan3_affine_dna4_simd<score>/8/real_time                     8264505 ns      8235000 ns           85 CUPS=2.81643G/s cells=23.2764M total=-6.08557M
seqan3_affine_dna4_simd<score>/16/real_time                    8787906 ns      8768622 ns           74 CUPS=2.64134G/s cells=23.2119M total=-5.30077M
seqan3_affine_dna4_simd<score>/24/real_time                    9933738 ns      9890493 ns           69 CUPS=2.32985G/s cells=23.1441M total=-4.98063M
seqan3_affine_dna4_simd<score>/32/real_time                   11355174 ns     11028946 ns           56 CUPS=2.03286G/s cells=23.0835M total=-4.07338M
seqan3_affine_dna4_simd<score>/40/real_time                   11108335 ns     11063083 ns           60 CUPS=2.07201G/s cells=23.0166M total=-4.42242M
seqan3_affine_dna4_simd<score>/48/real_time                   12102132 ns     12062811 ns           53 CUPS=1.8968G/s cells=22.9553M total=-3.96127M
seqan3_affine_dna4_simd<score>/56/real_time                   12775721 ns     12737870 ns           54 CUPS=1.79191G/s cells=22.8929M total=-4.10821M
seqan3_affine_dna4_simd<score>/64/real_time                   13963623 ns     13921420 ns           50 CUPS=1.63493G/s cells=22.8295M total=-3.89155M
seqan3_affine_dna4_simd<end_position>/0/real_time              7446336 ns      7434576 ns           92 CUPS=3.13553G/s cells=23.3482M total=-6.5469M
seqan3_affine_dna4_simd<end_position>/8/real_time              8261761 ns      8239475 ns           80 CUPS=2.81736G/s cells=23.2764M total=-5.7276M
seqan3_affine_dna4_simd<end_position>/16/real_time             8764267 ns      8740378 ns           74 CUPS=2.64847G/s cells=23.2119M total=-5.30077M
seqan3_affine_dna4_simd<end_position>/24/real_time             9604617 ns      9574521 ns           73 CUPS=2.40968G/s cells=23.1441M total=-5.26936M
seqan3_affine_dna4_simd<end_position>/32/real_time            10499088 ns     10469176 ns           68 CUPS=2.19862G/s cells=23.0835M total=-4.94625M
seqan3_affine_dna4_simd<end_position>/40/real_time            11236018 ns     11211016 ns           62 CUPS=2.04846G/s cells=23.0166M total=-4.56983M
seqan3_affine_dna4_simd<end_position>/48/real_time            12122554 ns     12097071 ns           56 CUPS=1.89361G/s cells=22.9553M total=-4.1855M
seqan3_affine_dna4_simd<end_position>/56/real_time            12678628 ns     12658943 ns           53 CUPS=1.80563G/s cells=22.8929M total=-4.03213M
seqan3_affine_dna4_simd<end_position>/64/real_time            13724687 ns     13691569 ns           51 CUPS=1.66339G/s cells=22.8295M total=-3.96938M
seqan3_affine_dna4_simd_parallel<score>/0/real_time            2087733 ns       483611 ns          352 CUPS=11.1835G/s cells=23.3482M total=-25.049M
seqan3_affine_dna4_simd_parallel<score>/8/real_time            2253220 ns       489483 ns          319 CUPS=10.3303G/s cells=23.2764M total=-22.8388M
seqan3_affine_dna4_simd_parallel<score>/16/real_time           2355413 ns       489446 ns          298 CUPS=9.85469G/s cells=23.2119M total=-21.3463M
seqan3_affine_dna4_simd_parallel<score>/24/real_time           3052400 ns       521400 ns          240 CUPS=7.58226G/s cells=23.1441M total=-17.3239M
seqan3_affine_dna4_simd_parallel<score>/32/real_time           2692046 ns       498588 ns          260 CUPS=8.57472G/s cells=23.0835M total=-18.9121M
seqan3_affine_dna4_simd_parallel<score>/40/real_time           2980305 ns       510594 ns          244 CUPS=7.72289G/s cells=23.0166M total=-17.9845M
seqan3_affine_dna4_simd_parallel<score>/48/real_time           3054948 ns       498150 ns          234 CUPS=7.51415G/s cells=22.9553M total=-17.4894M
seqan3_affine_dna4_simd_parallel<score>/56/real_time           3332705 ns       490600 ns          220 CUPS=6.86916G/s cells=22.8929M total=-16.7372M
seqan3_affine_dna4_simd_parallel<score>/64/real_time           3467255 ns       491353 ns          207 CUPS=6.58431G/s cells=22.8295M total=-16.111M
seqan3_affine_dna4_simd_parallel<end_position>/0/real_time     2091084 ns       490408 ns          333 CUPS=11.1656G/s cells=23.3482M total=-23.6969M
seqan3_affine_dna4_simd_parallel<end_position>/8/real_time     2246094 ns       488059 ns          320 CUPS=10.3631G/s cells=23.2764M total=-22.9104M
seqan3_affine_dna4_simd_parallel<end_position>/16/real_time    2428490 ns       493181 ns          271 CUPS=9.55814G/s cells=23.2119M total=-19.4123M
seqan3_affine_dna4_simd_parallel<end_position>/24/real_time    2543703 ns       482522 ns          276 CUPS=9.09858G/s cells=23.1441M total=-19.9225M
seqan3_affine_dna4_simd_parallel<end_position>/32/real_time    2757918 ns       494351 ns          242 CUPS=8.36991G/s cells=23.0835M total=-17.6028M
seqan3_affine_dna4_simd_parallel<end_position>/40/real_time    2933144 ns       481576 ns          238 CUPS=7.84706G/s cells=23.0166M total=-17.5423M
seqan3_affine_dna4_simd_parallel<end_position>/48/real_time    3147359 ns       490943 ns          211 CUPS=7.29352G/s cells=22.9553M total=-15.7704M
seqan3_affine_dna4_simd_parallel<end_position>/56/real_time    3436221 ns       505319 ns          207 CUPS=6.66223G/s cells=22.8929M total=-15.7481M
seqan3_affine_dna4_simd_parallel<end_position>/64/real_time    4233892 ns       531726 ns          168 CUPS=5.39208G/s cells=22.8295M total=-13.0756M
seqan2_affine_dna4_simd<score>/0/real_time                     6474490 ns      6433790 ns          100 CUPS=3.60619G/s cells=23.3482M total=-7.1162M
seqan2_affine_dna4_simd<score>/8/real_time                     9089863 ns      9065743 ns           74 CUPS=2.5607G/s cells=23.2764M total=-5.29803M
seqan2_affine_dna4_simd<score>/16/real_time                    9930999 ns      9902313 ns           67 CUPS=2.33731G/s cells=23.2119M total=-4.79934M
seqan2_affine_dna4_simd<score>/24/real_time                   10916453 ns     10864969 ns           64 CUPS=2.12011G/s cells=23.1441M total=-4.61971M
seqan2_affine_dna4_simd<score>/32/real_time                   11496536 ns     11466183 ns           60 CUPS=2.00787G/s cells=23.0835M total=-4.36434M
seqan2_affine_dna4_simd<score>/40/real_time                   12386365 ns     12343691 ns           55 CUPS=1.85822G/s cells=23.0166M total=-4.05389M
seqan2_affine_dna4_simd<score>/48/real_time                   15312298 ns     15100294 ns           51 CUPS=1.49914G/s cells=22.9553M total=-3.81179M
seqan2_affine_dna4_simd<score>/56/real_time                   14076291 ns     14029867 ns           45 CUPS=1.62634G/s cells=22.8929M total=-3.42351M
seqan2_affine_dna4_simd<score>/64/real_time                   14772515 ns     14730977 ns           43 CUPS=1.5454G/s cells=22.8295M total=-3.34673M
seqan2_affine_dna4_simd_parallel<score>/0/real_time            2170135 ns      1949401 ns          342 CUPS=10.7589G/s cells=23.3482M total=-24.3374M
seqan2_affine_dna4_simd_parallel<score>/8/real_time            2809440 ns      2642644 ns          180 CUPS=8.28506G/s cells=23.2764M total=-12.8871M
seqan2_affine_dna4_simd_parallel<score>/16/real_time           3046737 ns      2841892 ns          167 CUPS=7.6186G/s cells=23.2119M total=-11.9625M
seqan2_affine_dna4_simd_parallel<score>/24/real_time           4252225 ns      3561966 ns          119 CUPS=5.44282G/s cells=23.1441M total=-8.58978M
seqan2_affine_dna4_simd_parallel<score>/32/real_time           4735318 ns      3879006 ns          155 CUPS=4.87476G/s cells=23.0835M total=-11.2745M
seqan2_affine_dna4_simd_parallel<score>/40/real_time           4456511 ns      4060742 ns          178 CUPS=5.1647G/s cells=23.0166M total=-13.1198M
seqan2_affine_dna4_simd_parallel<score>/48/real_time           6129391 ns      5589870 ns          100 CUPS=3.74513G/s cells=22.9553M total=-7.4741M
seqan2_affine_dna4_simd_parallel<score>/56/real_time           9020744 ns      5767146 ns           82 CUPS=2.5378G/s cells=22.8929M total=-6.2384M
seqan2_affine_dna4_simd_parallel<score>/64/real_time           4532458 ns      4268313 ns          144 CUPS=5.03689G/s cells=22.8295M total=-11.2077M
```