@rrahn 2020-03-26
spurious coverage drop is not related to this change.